[#87] Changed scene skybox rotation into x-y-z format and fixed marker label setAttribute issue
