+++ app/imports/lib/aframeComponents/mouseoverEffectHyperlinkObject.js
@@ -8,12 +8,16 @@ AFRAME.registerComponent('mouseover-effect-hyperlink-object', {
     init() {
         this.el.addEventListener('mouseenter', () => {
             $('.marker-model', $(this.el))[0].setAttribute('model-opacity', 1);
-            $('.marker-label', $(this.el))[0].setAttribute('visible', 'true');
+            if ($('.marker-label', $(this.el))[0] !== undefined) {
+                $('.marker-label', $(this.el))[0].setAttribute('visible', 'true');
+            }
             $('.a-canvas').addClass('vw-mouseover');
         });
         this.el.addEventListener('mouseleave', () => {
             $('.marker-model', $(this.el))[0].setAttribute('model-opacity', 0.5);
-            $('.marker-label', $(this.el))[0].setAttribute('visible', 'false');
+            if ($('.marker-label', $(this.el))[0] !== undefined) {
+                $('.marker-label', $(this.el))[0].setAttribute('visible', 'false');
+            }
             $('.a-canvas').removeClass('vw-mouseover');
         });
     }
